# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 報修管理系統 API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RepairStatus = (typeof RepairStatus)[keyof typeof RepairStatus];

export const RepairStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export type RepairPriority =
  (typeof RepairPriority)[keyof typeof RepairPriority];

export const RepairPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Repair {
  id: number;
  title: string;
  location: string;
  category: string;
  status: RepairStatus;
  priority: RepairPriority;
  /** @nullable */
  description?: string | null;
  reportedBy: string;
  /** @nullable */
  assignedTo?: string | null;
  reportedAt: string;
  /** @nullable */
  resolvedAt?: string | null;
  /** @nullable */
  notes?: string | null;
}

export type RepairInputPriority =
  (typeof RepairInputPriority)[keyof typeof RepairInputPriority];

export const RepairInputPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface RepairInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  location: string;
  /** @minLength 1 */
  category: string;
  priority: RepairInputPriority;
  description?: string;
  /** @minLength 1 */
  reportedBy: string;
  assignedTo?: string;
  notes?: string;
}

export type RepairUpdateStatus =
  (typeof RepairUpdateStatus)[keyof typeof RepairUpdateStatus];

export const RepairUpdateStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export type RepairUpdatePriority =
  (typeof RepairUpdatePriority)[keyof typeof RepairUpdatePriority];

export const RepairUpdatePriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface RepairUpdate {
  /** @minLength 1 */
  title?: string;
  location?: string;
  category?: string;
  status?: RepairUpdateStatus;
  priority?: RepairUpdatePriority;
  description?: string;
  /** @minLength 1 */
  reportedBy?: string;
  assignedTo?: string;
  /** @nullable */
  resolvedAt?: string | null;
  notes?: string;
}

export type MonthlyStatsByStatus = {
  pending: number;
  in_progress: number;
  completed: number;
};

export type MonthlyStatsByCategory = { [key: string]: number };

export type MonthlyStatsByPriority = {
  low: number;
  medium: number;
  high: number;
};

export interface MonthlyStats {
  month: string;
  total: number;
  byStatus: MonthlyStatsByStatus;
  byCategory: MonthlyStatsByCategory;
  byPriority: MonthlyStatsByPriority;
  /** @nullable */
  avgResolutionDays?: number | null;
}

export interface Summary {
  total: number;
  pending: number;
  inProgress: number;
  completedThisMonth: number;
  highPriority: number;
}

export type CustomFieldFieldType =
  (typeof CustomFieldFieldType)[keyof typeof CustomFieldFieldType];

export const CustomFieldFieldType = {
  text: "text",
  number: "number",
  select: "select",
  date: "date",
} as const;

export interface CustomField {
  id: number;
  name: string;
  fieldType: CustomFieldFieldType;
  /** @nullable */
  options?: string[] | null;
  required: boolean;
  sortOrder: number;
}

export type CustomFieldInputFieldType =
  (typeof CustomFieldInputFieldType)[keyof typeof CustomFieldInputFieldType];

export const CustomFieldInputFieldType = {
  text: "text",
  number: "number",
  select: "select",
  date: "date",
} as const;

export interface CustomFieldInput {
  /** @minLength 1 */
  name: string;
  fieldType: CustomFieldInputFieldType;
  options?: string[];
  required?: boolean;
}

export interface CustomFieldUpdate {
  /** @minLength 1 */
  name?: string;
  options?: string[];
  required?: boolean;
  sortOrder?: number;
}

export interface CustomValue {
  id: number;
  repairId: number;
  fieldId: number;
  value: string;
}

export interface CustomValueInput {
  fieldId: number;
  value: string;
}

export interface CustomValuesPayload {
  values: CustomValueInput[];
}

export interface Category {
  id: number;
  slug: string;
  label: string;
  color: string;
  sortOrder: number;
}

export interface CategoryInput {
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  label: string;
  color?: string;
  sortOrder?: number;
}

export interface CategoryUpdate {
  /** @minLength 1 */
  label?: string;
  color?: string;
  sortOrder?: number;
}

export interface TrendPoint {
  month: string;
  total: number;
  completed: number;
  pending: number;
}

export type ListRepairsParams = {
  status?: ListRepairsStatus;
  category?: string;
  priority?: ListRepairsPriority;
  /**
   * Filter by month (YYYY-MM format)
   */
  month?: string;
};

export type ListRepairsStatus =
  (typeof ListRepairsStatus)[keyof typeof ListRepairsStatus];

export const ListRepairsStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export type ListRepairsPriority =
  (typeof ListRepairsPriority)[keyof typeof ListRepairsPriority];

export const ListRepairsPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export type GetMonthlyStatsParams = {
  /**
   * Month in YYYY-MM format (defaults to current month)
   */
  month?: string;
};