# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: 報修管理系統 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /custom-fields:
    get:
      operationId: listCustomFields
      tags: [customFields]
      summary: List all custom field definitions
      responses:
        "200":
          description: List of custom field definitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomField"
    post:
      operationId: createCustomField
      tags: [customFields]
      summary: Create a custom field definition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomFieldInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomField"
        "400":
          description: Bad request
  /custom-fields/{id}:
    patch:
      operationId: updateCustomField
      tags: [customFields]
      summary: Update a custom field definition
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomFieldUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomField"
        "404":
          description: Not found
    delete:
      operationId: deleteCustomField
      tags: [customFields]
      summary: Delete a custom field definition
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /repairs/{id}/custom-values:
    put:
      operationId: setRepairCustomValues
      tags: [repairs]
      summary: Set custom field values for a repair
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomValuesPayload"
      responses:
        "200":
          description: Updated custom values
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomValue"
        "404":
          description: Repair not found
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Bad request
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  schemas:
...
    CustomField:
      type: object
      required: [id, name, fieldType, required, sortOrder]
      properties:
        id:
          type: integer
        name:
          type: string
        fieldType:
          type: string
          enum: [text, number, select, date]
        options:
          type: ["array", "null"]
          items:
            type: string
        required:
          type: boolean
        sortOrder:
          type: integer
    CustomFieldInput:
      type: object
      required: [name, fieldType]
      properties:
        name:
          type: string
          minLength: 1
        fieldType:
          type: string
          enum: [text, number, select, date]
        options:
          type: array
          items:
            type: string
        required:
          type: boolean
    CustomFieldUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        options:
          type: array
          items:
            type: string
        required:
          type: boolean
        sortOrder:
          type: integer
    CustomValue:
      type: object
      required: [id, repairId, fieldId, value]
      properties:
        id:
          type: integer
        repairId:
          type: integer
        fieldId:
          type: integer
        value:
          type: string
    CustomValueInput:
      type: object
      required: [fieldId, value]
      properties:
        fieldId:
          type: integer
        value:
          type: string
    CustomValuesPayload:
      type: object
      required: [values]
      properties:
        values:
          type: array
          items:
            $ref: "#/components/schemas/CustomValueInput"
    Category:
      type: object
      required: [id, slug, label, color, sortOrder]
      properties:
        id:
          type: integer
        slug:
          type: string
        label:
          type: string
        color:
          type: string
        sortOrder:
          type: integer
    CategoryInput:
      type: object
      required: [slug, label]
      properties:
        slug:
          type: string
          minLength: 1
        label:
          type: string
          minLength: 1
        color:
          type: string
        sortOrder:
          type: integer
    CategoryUpdate:
      type: object
      properties:
        label:
          type: string
          minLength: 1
        color:
          type: string
        sortOrder:
          type: integer
...