# lib/api-zod/src/generated/types/customField.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 報修管理系統 API
 * OpenAPI spec version: 0.1.0
 */
import type { CustomFieldFieldType } from "./customFieldFieldType";

export interface CustomField {
  id: number;
  name: string;
  fieldType: CustomFieldFieldType;
  /** @nullable */
  options?: string[] | null;
  required: boolean;
  sortOrder: number;
}