# lib/api-zod/src/generated/types/customFieldInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 報修管理系統 API
 * OpenAPI spec version: 0.1.0
 */
import type { CustomFieldInputFieldType } from "./customFieldInputFieldType";

export interface CustomFieldInput {
  /** @minLength 1 */
  name: string;
  fieldType: CustomFieldInputFieldType;
  options?: string[];
  required?: boolean;
}