# lib/api-zod/src/generated/types/customFieldUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 報修管理系統 API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomFieldUpdate {
  /** @minLength 1 */
  name?: string;
  options?: string[];
  required?: boolean;
  sortOrder?: number;
}