# lib/api-zod/src/generated/types/customValuesPayload.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 報修管理系統 API
 * OpenAPI spec version: 0.1.0
 */
import type { CustomValueInput } from "./customValueInput";

export interface CustomValuesPayload {
  values: CustomValueInput[];
}